refactor: extract user list component without testing
idk if it works :/
