+++ src/components/user-list.tsx
@@ -1,0 +1,11 @@
+import type { ReactElement } from "react";
+
+import { Stack, type StackProps } from "@mui/material";
+
+import type { UserItem } from "@/components/user-item";
+
+export type UserListProps = Omit<StackProps, "children"> & {
+  children: ReactElement<typeof UserItem>[];
+};
+
+export const UserList = (props: UserListProps) => <Stack gap={3} {...props} />;
